+++ jib-maven-plugin/CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Changed the default base image from the `adoptopenjdk` images to the `eclipse-temurin` images on Docker Hub. Note that Temurin (by Adoptium) is the new name of AdoptOpenJDK. ([#xyz](https://github.com/GoogleContainerTools/jib/pull/xyz))
+
 ### Fixed
 
 ## 3.1.4